Stop using mapi.Connection.cmd in the profiler connection
Connecting to the profiler should not be going through
`mapi.Connection.cmd`.

The issue is that if the server is running a query, as soon as the
profiler connection is established a profiler object is sent to the
client. `mapi.Connection.cmd` reads a block and tries to interpret this
data according to the MAPI protocol and fails, since the JSON object
does not conform to the MAPI specification.

The solution is to use `mapi.Connection._putblock` and
`mapi.Connection._getblock` to implement a profiler appropriate
connection handshake.